// packages/b2c-tooling-sdk/src/clients/custom-apis.generated.ts
export interface components {
    schemas: {
// ...
        CustomApiEndpoint: {
            /** @example loyalty-info */
            apiName?: string;
            /** @example v1 */
            apiVersion?: string;
            /** @example test_bc_wapi */
            cartridgeName?: string;
            /** @example /customers */
            endpointPath?: string;
            /** @example API schema not found. */
            errorReason?: string;
            /**
             * @example GET
             * @enum {string}
             */
            httpMethod?: "GET" | "POST" | "PUT" | "DELETE" | "PATCH" | "OPTIONS" | "HEAD";
            /** @example 10bd7f2dc40ab7aede7f0d60e5c3a783 */
            id?: string;
            /** @example script.js */
            implementationScript?: string;
            /** @example getLoyaltyInfo */
            operationId?: string;
            /**
             * @example ShopperToken
             * @enum {string}
             */
            securityScheme?: "ShopperToken" | "AmOAuth2";
            /** @example schema.yaml */
            schemaFile?: string;
            siteId?: components["schemas"]["SiteId"];
            status?: components["schemas"]["EndpointStatus"];
        };
// ...
    };
    responses: never;
    parameters: {
        /**
         * @description An identifier for the organization the request is being made by
         * @example f_ecom_zzxy_prd
         */
        organizationId: components["schemas"]["OrganizationId"];
        status: components["schemas"]["EndpointStatus"];
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}